tools/scripts/requirements.txt
openpyxl>=3.1.2
psycopg2-binary>=2.9.9
